Misc: agent-inferlet fix-quality prompts, driver fixes, terminalbench scripts
- inferlets/openhands-agent: root-cause-first prompt rules, empty-finish
  retry cap, condensation cooldown/target, per-step metrics.
- client.py: propagate websocket disconnect to per-process event queues so
  awaiting callers fail fast instead of hanging.
- driver/dev batching: int64 cumulative bit offsets (int32 overflowed on
  large batches); driver/vllm config: optional max_model_len override to
  shrink KV pre-allocation; uv.lock for the vllm driver env.
- NEWCOMER.md: design-principles / first-inferlet guide.
- integrations/terminalbench: Terminal-Bench run scripts (pie + baseline).

+++ driver/dev/src/pie_driver_dev/batching.py
@@ -141,7 +141,7 @@ class Batch:
         if self.single_token_mode:
             self.attention_masks = None
         else:
-            token_acc_seq_lens_np = np.zeros(self.total_tokens + 1, dtype=np.int32)
+            token_acc_seq_lens_np = np.zeros(self.total_tokens + 1, dtype=np.int64)
             np.cumsum(all_seq_lens, out=token_acc_seq_lens_np[1:])
             self.attention_masks = decode_brle_batch(
                 flattened_masks_u32, mask_indptr, position_ids_np, token_acc_seq_lens_np
@@ -1040,7 +1040,7 @@ def decode_brle_batch(
         flattened_masks: Concatenated BRLE run lengths (int32)
         mask_indptr: Pointers to BRLE ranges per token (int32)
         position_ids: Position of each token, defines valid_len (int32)
-        token_acc_seq_lens: Cumulative bit offsets per token (int32)
+        token_acc_seq_lens: Cumulative bit offsets per token (int64)
 
     Returns:
         Flat boolean array with all mask values
